LSP should not error when flowconfig has warnings
Summary:
when flowconfig contains invalid options (usually, deprecated ones that haven't been removed yet) the flowconfig parser returns warnings. by default, `CommandUtils.read_config_and_exit` treats them as errors and exits.

this is bad for `flow lsp` because it appears to the editor like it just crashed. in vscode, it quickly retries it 5 times and disables the extension until you restart vscode!

this diff makes `flow lsp` not error on warnings.

in the future, we should send the warnings to the client with `window/showMessage` instead of logging them to stderr. in fact, we should do the same for errors instead of exiting! we could also use `showMessageRequest` to include an "Open .flowconfig" button.

Reviewed By: dsainati1

Differential Revision: D28571084

fbshipit-source-id: b7f0177be85f5a8d144ef28d8bcd1d5c0ed20b37