simplify Watchman.get_mergebase_and_changes
Summary:
when we receive an `hg.update` notification from our persistent watchman connection, we call `Watchman.get_mergebase_and_changes` to get the changes, using a separate connection.

we don't need to use `call_on_instance`. `call_on_instance` only does something if the persistent connection has died (it re-sets up the watch, subscribes, and reopens the persistent connection), but the connection could not have died because we call this in response to a successful reply from watchman and there's no gap in between for it to die.

removing it simplifies `get_mergebase_and_changes`, and makes `call_on_instance` _only_ used for the persistent connection for which it was designed, unlocking further improvements.

Reviewed By: samwgoldman

Differential Revision: D28562314

fbshipit-source-id: 935fea2f75b73338aa93649d0eaac03e50af2913