make Watchman.get_changes retry internally
Summary:
`Watchman.get_changes` is called in an infinite loop in `FileWatcher.WatchmanListenLoop`.

before: if it fails we return a `Watchman_dead`, the loop sleeps for 1 second, and then next iteration of the loop reconnects *if the backoff delay has elapsed*, otherwise it just returns `Watchman_dead` again, the loop sleeps for 1 second again, etc.

after: `get_changes` retries internally, using a single `Lwt_unix.sleep` for the backoff delay instead of checking again every second. this means it only returns to the caller on success, or on a fatal error (like a restart, or out of retries).

I also resolved a long-standing TODO comment: "Should we die if we get this for too long?" yes! we already die if we fail to reconnect 8 times in a row. now we also die if we reconnect, get an error, reconnect, get an error, and reconnect then get an error (3x reconnect + error) without any successful change events in between.

Reviewed By: nmote

Differential Revision: D28562313

fbshipit-source-id: 3d105f51945cd5aab3f11f70ae1ad3bcea32bec6