Model refinements over previous refinements instead of total refinements
Summary:
This diff restructures the internal representation of refinements. Instead of tracking the total refinements over an SSA id, we instead track the _incremental_ refinements over the previous refinement.

## Refinement Ids and SSA Ids

We introduce a new concept of "refinement ids" in this diff, which are keys into a heap of refinements. We maintain a mapping from ssa ids -> refinement ids to track the latest refinement over an ssa_id.

We keep a mapping from ssa ids to refinement ids in a stack representing the "last refinement over this ssa in the current scope"

This stack operates in almost the exact same way that the old refinement stack prior to this diff worked.

## Refinement chains

Instead of storing refinements directly in the heap, we store `refinement_chains`. These chains allow us to build up the full structure of the total refinement but represent only the incremental updates to a refinement over the previous version.

A critical invariant that we maintain is that a refinement from one refinement scope will not have a reference to a refinement from another refinement scope within it. This keeps it easy for us to do things like negate new refinements, which is a key feature of the EnvBuilder.

Reviewed By: mvitousek

Differential Revision: D27775784

fbshipit-source-id: b2730571ad6d66b82bcc036597bc0e51a2a049c9