flow fix command
Summary:
Introduces a `flow fix` command which applies the "replace object with interface" and "replace method with arrow" quickfixes to an entire codebase. This codemod will also automatically pick up on any new quickfixes we introduce into this machinery, so this gives us a foundation for expressing language changes as codemods and then automatically applying them to the codebase.

Right now, the decision of which quickfixes do and don't get applied in this codemod is a matter of the abstraction I carved out in `code_action_service.ml`. There's a new function `ast_transform_of_error` which is what the codemod calls. Other quickfixes, like "did you mean", are not included in the codemod because they're not expressed as ast transforms.

Reviewed By: mroch

Differential Revision: D26785767

fbshipit-source-id: 87653b9c41bc02b8affc670768052813c08dcaed