Upper-bound autocomplete for object keys
Summary:
Autocomplete for object keys. When autocomplete is requested in an object key position, we compute the upper bound of the object and use `members_of_type` to get the key names to suggest in autocomplete. To ensure that we're able to compute an upper bound for the object, the object key hook makes object literal's type always an OpenT. For object key names that don't parse as identifiers, we suggest string literal keys. We use the lower-bound object type to compute what object keys not to suggest, i.e. the ones that are already included in the object, similarly to what we do for JSX attribute autocomplete. Unlike in JSX attribute autocomplete, we use the lower bound type rather than computing the already-present object keys from the AST; using the lower bound type is preferable in this case because the type knows about keys from spreads; for example see the `upper_bounds_10.js` test.

Fixes: https://github.com/facebook/flow/issues/7723

Reviewed By: mroch

Differential Revision: D28445414

fbshipit-source-id: b1fbddd57767e0622c0e7e1bcf2fad8829a501c3