Test for an incremental mode issue introduced by sighash loading
Summary:
I realized yesterday that a specific, and likely rare, series of events could lead Flow to miss type errors that it should catch after sighashes are loaded from saved state.

The issue lies in how Flow treats changes to the repository between when the saved state was generated and the current repository state, when a lazy server is started. We consider two distinct sets of changes:
1. Those made between when the saved state was generated and the current merge base with the main branch (changes already committed). These are required to appear in the `.flow.saved_state_file_changes` file.
2. Those made after the current merge base (your local changes).

Currently, when loading from a saved state, Flow reparses all of the files in (1) but not their dependencies or dependents, and uses that to update the data that was loaded from saved state (most notably the dependency graph).

Then, Flow kicks off a recheck with the files in (2) as inputs, checking all of their dependents.

We don't recheck the files in (1) because we don't need to in order to satisfy the guarantee of lazy mode: that we'll find errors caused by your local changes. We only do the bare minimum to update the information that we loaded from saved state.

Unfortunately, with the introduction of sighash loading, this bare minimum is no longer enough.

Here's what's happening in this test:

1. We start Flow and create a saved state
2. We modify A.js and B.js (this is to simulate a committed change to the main branch)
3. We start Flow in lazy mode, and it reparses both A.js and B.js but does not recheck them
4. We revert A.js to the state that it was in in step (1).
5. Flow does a recheck, but sees that the current sighash of A.js matches the one that it loaded from the saved state, so it skips checking B.js, thereby missing the error that was introduced.

I'd be surprised if this happens much, if ever, in practice, because it can only happen if a file is changed, that change is committed, an updated saved state is not generated, and then someone locally reverts the change to that file. Still, it's a risk, and it should probably be addressed.

Note that simply declining to load the sighashes for changed files is not sufficient to address this issue. We could construct an example where the changes to a file affect the sighash of its dependents in a similar way. I believe we would need to decline to load sighashes for the changed files and all of their recursive dependents, though it's not yet clear to me what we would need to do to account for changes in the dependency graph.

Reviewed By: mroch

Differential Revision: D28715190

fbshipit-source-id: 63b2b073364af947889f5203ab1368b033223ebc