survive most watchman restarts
Summary:
when Watchman restarts, we miss any filesystem changes that might happen while it's down. this can cause our internal state to be out of date.

we currently take the nuclear option of completely restarting the server, since that clears out _all_ the state. this is bad for UX because all IDE services stop working.

in watchman lazy mode, we only "focus" files that have changed since mergebase -- that is, we assume `master` is clean and only check files you've changed locally. watchman can tell us which files were changed since mergebase.

to safely survive a watchman restart, we have to figure out which files could possibly have changed while watchman wasn't watching and recheck all of them. there are a few possibilities:

1) a file that was previously unchanged since mergebase, could become changed
2) a changed file could change again
3) a changed file could be reverted (become unchanged)
4) the mergebase could change, causing some "unfocused" files to change

if we ask watchman for all of the changes since mergebase when it restarts, it will tell us about cases (1) and (2), but it won't know about (3). it will tell us if the mergebase changed (4), but not which unfocused files changed. for case (3), we already know which files were previously changed: the focused files.

so we can overapproximate the potentially changed files as the changes since the mergebase, plus all the previously focused files, AS LONG AS the mergebase didn't change. so, we will continue to restart if the mergebase changed, but can survive cases (1) - (3).

NOTE: if the mergebase changed, in theory we could ask `hg` for the files that changed between the old and new mergebases, but that seems rare and is left for future work.

when we recheck all of these files, the file hashes and signature hashes prevent us from doing any unnecessary work. IDE services stay working the whole time (modulo potentially being stale since rechecks aren't triggered, but that's already the case while rechecks are in progress)

Reviewed By: nmote

Differential Revision: D28715425

fbshipit-source-id: 06b1918764a514ad62fda43fcbda9fc33802e3fd