Enable new check mode by default
Summary:
This option enables a new architecture for Flow which uses significantly less
RAM for large projects. We have been using this internally for a few weeks and
have not noticed any issues.

For now, people can opt out of this mode by adding the following line to their
project's .flowconfig:

> experimental.new_check = false

In an upcoming version of Flow, we will remove the configuration, leaving "new
check" mode as the only available mode.

Reviewed By: panagosg7

Differential Revision: D28810669

fbshipit-source-id: 23f6cb489e098366dac878350b91bfb0a48c3c7d