[lsp] Quickfix for optional chaining
Summary:
This diff implements a quickfix for adding optional chaining when we are accessing an object that is possibly nullable.

It will match on the error of `IncompatibleUse` with `upper_kind = Error_message.IncompatibleGetPropT`, and the lower reason is a nullable type. This should corresponds to the case of `potentiallyNullableObject.someProperty`, which we wants to transform into `potentiallyNullableObject?.someProperty`.

This diff is still using the AST transformation approach, so that we can automatically handle the dot syntax case and bracket syntax case.

Reviewed By: vrama628

Differential Revision: D28887044

fbshipit-source-id: 4ba337d3ec6d5a9e0a3fbff3e7af17e1ca11fbd6