[providers] Fix to providers of loop declarations
Summary: This diff fixes a bug with providers for variables declared in loops. It also fixes a bug in the `all_providers` function of provider_api, which was needed in order to test this change, and it introduces a distinction between `Initialized` and `Annotated` variables within the provider analysis--this was originally built as part of D28145236, but the distinction is relevant here and including it in this diff avoids needing to remember to make the corresponding change as part of that diff.

Reviewed By: jbrown215

Differential Revision: D28913804

fbshipit-source-id: f6547bf786ce37b49b39bce93d8843fd5586ab01