[new-env] Add a mode where cycles are detected and errored on, but reordering doesn't actually happen
Summary: It's likely that we'll want to signal errors on dependency cycles in order to power codemods etc before we actually can check programs in reordered form. This diff adds a mode for doing just that--it uses the reordering analysis to check for dependency cycles and raises errors on them, but lexical order is used in actual checking.

Reviewed By: panagosg7

Differential Revision: D28074893

fbshipit-source-id: 3fd520ddb12b1d40870487932ed698abc458c156