[lsp][extract] Collect a list of statements AST for function extraction
Summary:
As the first step of implementing function extraction refactor, we must first figure out what to extract. More precisely, we should solve the problem that:

> Given a location range, find a list of statements that should be part of the new extracted function.

To implement the selection, I used a visitor over the AST that collects the statements into a mutable list. The traversal does not care about the kind of the AST node, but only cares about the relation between the statement range and the selection range.

There are only 4 possible relations between these two ranges:

1. selection range is completely contained inside statement range.
In this case, the statement should not be collected, because if we collect this then we will only end up collecting all top-level statements. However, we should recurse down to make the traversal find smaller statements.
2. statement range is completely contained inside selection range.
In this case, we should collect the statement, since it's within the selection. However, we should not recurse down, because otherwise we might collect both a block statement and smaller statements nested inside the block.
3. disjoint
Obviously we do not collect and do not recurse down.
4. Intersection
In this case, we also should not collect, since the user does not select the statement in full. We should also not recurse down, because otherwise we might end up partially collecting a block. To see why that's the case, we can look at an example below:

```
// selection start
const a = 3;
{
  foo();
  // selection end
  bar();
}
```

If we do recurse down, the statements we collect forms

```
const a = 3;
foo();
```

On its own it does not look too bad, but when we are replacing this range with a `newFunction()` call, the original code will become bad:

```
// selection start
newFunction();
  bar();
}
```

Therefore, we must not recurse down and must completely drop the block from considerartion.

Reviewed By: vrama628

Differential Revision: D28940954

fbshipit-source-id: 0eed1c6638349fab740933bf57e30db7c1b6d6f1