separate function `this` types into two views: one for calls and one for subtyping
Summary:
The new model we are using for the `this` parameter types of class methods requires us to be able to ignore the `this` parameter when subtyping classes. This is easy when simply verifying the class hierarchy, as we have the entire structure of the class available when doing this, and can simply choose to ignore the `this` parameters of any methods we encounter. It is more difficult if we are doing structural checks against interfaces, since we do not have the actual method types available at this point anymore.

Instead, we create method types with two `this` parameters, one that gets used when calling the method to ensure that its `this` argument is compatible with this parameter. The other, in later diffs, will be used for subtyping and for class/interface methods will be `mixed` to allow these to more easily subtype each other.

The two cases where using `mixed` for class method subtyping might be unsound are when class methods are being compared against toplevel functions or object methods. Both of these are only possible if the class method has been unbound or if the class is being subtyped with an object, both of which are now banned earlier in this stack.

Reviewed By: mvitousek

Differential Revision: D26489429

fbshipit-source-id: 98e96e7a7172fbed6fb5da088ce65a83cbcd8ce4