update core libs to use this annotations for thisArg callbacks
Summary:
Previously, the `this` type of the callbacks in library functions like `Array.map` was unchecked, and the `thisArg` argument to this function was simply mixed. We did not enforce that uses of `this` inside the callback were compatible with the `thisArg`. With the new `this` typing improvements, we can actually ensure that callers of `map` and similar functions use `this` properly inside their callbacks.

This means that code like the following:

```
class A {
    y = 3;
    x = [1,2,3].map( function(x) { this.y + x })
}
```

will now properly error. To make this work, the user should either use an autobinding arrow function for the callback, or pass the correct `thisArg` to the function (in this case `this`).

This also enables `this` annotations by default, since if they are going to be part of our core libraries they are going to be required.

Reviewed By: jbrown215

Differential Revision: D27742331

fbshipit-source-id: 0a7c61f9dfa36115b4e64b7b3aeaed4d365ee65b