infer mixed for interface methods
Summary:
Now that interface methods cannot be unbound, it is safe to infer `mixed` for their `this` parameter, as they have no body to check where this may pose an issue.

In order to make this work, this diff also includes the logic to ignore the `this` parameters of class methods when doing function-function subtyping, as this would otherwise cause spurious errors with structural subtyping.

Reviewed By: mvitousek

Differential Revision: D26489428

fbshipit-source-id: e888cfc1564e4cd2a2f4c2f715a2f02a99011454