[lsp][extract] Create AST for new functions that contains extracted statements
Summary:
This diff creates an AST for the extracted `newFunction`.

Right now, it does very naive function creation: it does not account for locally undefined variables that needs to be passed in as parameters, and it does not returns variables created/modified in the statements.

Since this diff is just for basic setup, above concerns will be address in future diffs.

Reviewed By: vrama628

Differential Revision: D28947074

fbshipit-source-id: 9856040c2ac161a2abc8174f67f9495d1677d0fa