[lsp][extract] Check whether user's selection range is valid for function extraction
Summary:
When user selects something and wants to extract them to function, not all selection range makes sense. For selection ranges that are ambiguous or invalid, we should not display an option for extracting function refactor.

This diff implements the TypeScript behavior: user selection must only contain full statements, potentially with some whitespace before and after. Since we are operating on strings, this behavior translates to the following criterion in terms of AST location:

> union of all selected statements' location equals the union of all touched statements' location.

, where "touch a statement" means that the selection range intersects with a statement.

For the implementation, I created another visitor that unions together locations using the above criterion, and then compare the location with the union of all statement locations.

Reviewed By: vrama628

Differential Revision: D28950089

fbshipit-source-id: 7e3a317e253f38a207b928b20bf87287c88d92ca