[lsp][extract] Implement transform that replaces extracted statements with `newFunction()` call
Summary:
This diff adds a transformation that replaces the extracted statements with a `newFunction()` call.

It is implemented by replacing the first statement with `newFunction()` and the rest of the statement with nothing. This requires a check that the statement list is non-empty, which conveniently also serves as a guard for empty extracted statements when we connected it to the rest of the LSP service.

Note that we cannot replace the rest of the statements with `Statement.Empty`, because otherwise `;` will be inserted in the places of the rest statements. We also need to give the inserted the `newFunction()` a location of the union of all selected statements, since otherwise pretty printer will insert an empty line after `newFunction()`.

Reviewed By: vrama628

Differential Revision: D28977422

fbshipit-source-id: e5afd0dcdea51374a05e7b1a118b693e1b65aa83