[lsp][extract][easy] Fix a bug where exact selection of single statement doesn't extract any statement
Summary:
Currently there is an edge case that is unaccounted. If the selection range is exactly equal to the range of a statement, then the check `Loc.contains extract_range statement_loc` will also be true, which causes it to recurse down and end up collecting nothing.

Mathematically, a statement should be collected if `extract_range >= statement_loc`, and we should recurse down if `statement_loc > extract_range`. The issue of the current code is that `Loc.contains` represents the `>=` operator instead of `>`, which causes the bug at the equal case.

The fix is simple, we flip the order of the first if and the else-if case, so that the equal case is checked first.

Reviewed By: vrama628

Differential Revision: D28973652

fbshipit-source-id: da0ed61f97a2b464ddd5fcb20ae9c1f8cae6c165