[lsp][extract] Connect function extraction refactor module to code action service
Summary:
This diff connects several implemented functions together so that we can play test a very incomplete version of function extraction code action.

Right now, it can only replace the extracted statements with `newFunction()`. Support for inserting the definition of `newFunction` to the top-level will be added in a followup diff.

Reviewed By: vrama628

Differential Revision: D28978129

fbshipit-source-id: e862a0b09cdf5bae5d21195da87316b9587d4787