truncate speculation branches
Summary:
For programs of the form
```
type T =
  | 'a_1'
  | ...
  | 'a_n';

declare var x: string;
(x: T); // error
```
flow would print an error that looked like
```
Cannot cast `x` to `T` because: [incompatible-cast]
 - Either string [1] is incompatible with string literal `a_1` [2].
 - Or string [1] is incompatible with string literal `a_2` [3].
 - ...
 - Or string [1] is incompatible with string literal `a_n` [n+1].

References:
...
```
For large enough n, printing out these errors could get really slow.

This diff changes the default behavior of printing of speculation errors, so that it prints
up to 50 branches by default and then includes an entry
```
 - Only showing the the first 50 branches. To see all branches, re-run Flow with --show-all-branches.
```
that instructs the user to pass in the `--show-all-branches` flag to print the rest of the branches.

Note that this flag is already used for printing branches that had been deemed "irrelevant"
for the default behavior. Ideally we could use a different flag, but the wording fit really well
here.

Reviewed By: samwgoldman

Differential Revision: D28982392

fbshipit-source-id: 0b4754c9a39259715e9a95b0a23f22af0024eb4a