remove experimental this annot option from configs
Summary: Now that we are using `this` annotations inside our core libraries, it no longer makes sense to allow users to disable `this` annotations, since this would generate errors in the core libraries every check. This removes the config option that was previously experimental, and removes the error message associated with using the annotations when the option was disabled.

Reviewed By: jbrown215

Differential Revision: D28967317

fbshipit-source-id: 4858d269a1c1e648b44d008913513a4464ea6ccb