[enums] Enum method this typing
Summary:
`this` is used internally in the Flow Enums runtime, and should be the enum object. We can enforce this by updating the type definitions here.

Note that we don't actually use `$EnumProto` directly in Flow, as enums aren't objects, but rather when we call one of the "enum methods" we access them off of this type definition. We can't make this an interface as the enum object isn't actually a subtype of `$EnumProto`.

Reviewed By: dsainati1

Differential Revision: D28982976

fbshipit-source-id: 264cad96372ef8e0bcb87c91c3a4db24a061a90c