[lsp][extract] Allow refactor when user exactly selects a single statement
Summary:
D28973652 (https://github.com/facebook/flow/commit/1628d9fdd9017bd7301c4c63da80757694687322) fixes the bug when the user exactly selects a single statement but the statement is not extracted. This diff fixes a similar issue in the `allow_refactor_extraction` function.

Here, when the selection range is equal to the statement range, the `contains` check will still pass and it will recurse down and end up doing nothing. Therefore, we need to add an equal case check to prevent that. This change effectively turns the guard into `statement_loc > extract_range`, which is exactly what we want.

Reviewed By: vrama628

Differential Revision: D29033906

fbshipit-source-id: c4281135b04cda6ac90c0532658efb722844a04c