[lsp][extract] Support inserting extracted function at the end of the program
Summary:
This diff completes the MVP implementation of extract to function refactor. More concretely, if the selected statements defines no new variable, uses no variables from the outside, then the current implementation is good enough. The implementation will adds the extracted function to the end of the program, which matches the current TypeScript language service behavior.

Unfortunately, the diff given to LSP is a little cluttered, since the insertion diff will try to insert the statement to the end of program without any newline or space. Fixing the pretty-printing will be the focus of another diff.

Reviewed By: vrama628

Differential Revision: D29036263

fbshipit-source-id: f8499ba80ce06c27b68793c44fa235b355546c3d