[lsp][extract] Combine `extract_statements` and `allow_refactor_extraction`
Summary:
This diff aims to slightly clear up the logic and improve the performance of extract function code action.

Currently we are making two passes over the AST to decide whether extraction should be allowed: the first pass collects the statements to extract and the second pass computes the union of touched locations. When we use the check `union of touched locations == union of selected statements` to decide whether the selection contains partially selected statements, which should not be allowed.

This two-pass approach is obviously less efficient than a one-pass approach. In addition, it is also unnecessary. All we need is to find whether there are partial selection of statements, which we can easily find simply by `Loc.intersects extract_range statement_loc` during the AST pass (assuming the contains case has been already checked, which is the case of the existing code).

Therefore, in this diff, I merged the two visitor into one and make the `extract_statements` function return `None` to convey that extraction is not allowed.

In the future, this visitor can be used to collect even more information during the visit: e.g. whether there is an await/yield, newly defined variables and used variables, etc

Reviewed By: vrama628

Differential Revision: D29041210

fbshipit-source-id: ae55ca4f6d88b8c207b7e28094706effa746af1f