[ast-differ] Coalesce `Replace` and `Delete` AST diff into a single Replace diff spanning both replace and delete
Summary:
This diff adds a special case to `edits_of_changes` similar to the existing insert after replace case. In this case, we will coalesce a replace and delete into a single replace spanning both, given that the `new_statement`'s location is a union of both.

For example:
`replace(L1, 'foo();', newFunction()), delete(L2-L3, 'bar(); baz();')` will be come a single `replace(L1-L3, 'foo();', newFunction())`, if `newFunction()` has a loc of `L1-L3`.

This is intended to support pretty-printing the diff without an extra empty line when we are replacing a list of consecutive statements into one statement. (Since the default behavior of `Delete` diff printing is to replace the original text with empty string) As far as I know, this should only affect the extract function refactor, which I specifically make the `newFunction()` call AST spans the location of all the original locations instead of giving it a dummy one.

Reviewed By: vrama628

Differential Revision: D29002810

fbshipit-source-id: 938d46b2dc59119ddc6e978daffd344f0abe5471