fix error positioning issue with type application use_ops
Summary:
The example program added to the test below was causing a `prop-missing` type error to be located on the definition of `forEach` in `Map`. See https://flow.org/try/#0CYUwxgNghgTiAEkoGdnwLJQA4B4BqAfPAN4BQ8F8AZgPYwCiUYAFjgBoEAUYUEEARkwDWVAHYAueJwC22SZlyEAlPAC8RaQEsAHiGBLJANxqbgAbnKUA5iAAunA-DzwAPvGOmLleMjudDvACuIJJ4jgr4BBYAvqRxATDwVjA0gVh6ACJQtlBq8KIgAO4Y2A4WpAlJKWl5yanpwFk5AHQ29krldWnNWIHIzJwARIMdcaRdDU1QzbQMTAMy2CrqJJYUsljNvvYADKPRo2OkQA in version 0.152.0 to see what this looks like.

In practice, since this definition of `Map` is pulled from the libdefs, this can result in a type error that is located in library definitions, which will can crash Flow's error printer.

This is a result of some changes that were made to `flow_use_op` as part of the `134` release; which standardized how `use_op`s were selected to no longer be dependent on control flow. However, in this instance the logic for `pick_use_op` resulted in us never replacing `TypeApplication` `use_op`s, which should never be primary locations for errors when they result from unification. This adds another override so that the `FunCall` use_op can take precedence over the instantiation in this example.

Reviewed By: jbrown215

Differential Revision: D29041393

fbshipit-source-id: 9ae14d37a7ec973456e54c7973932a0a120ee839