Add type sig to parse artifacts cache
Summary:
The parse artifacts cache is used as input to typecheck_contents. In a
follow-up, I am going to add type_sig as an input to that function. We already
produce a type sig during parsing. This change simply adds it to the cache, so
we can pass it along.

My immediate goal is to produce the exported_locals environment from type sigs
instead of file sigs. Longer term, my aim is to replace file sigs with type sigs
entirely, since they have a lot of overlap.

Reviewed By: dsainati1

Differential Revision: D29033872

fbshipit-source-id: eab06514320b76615d991c2b6e46e0d13f0c6485