[lsp][extract] Generate extract function refactors for every named function that contains extracted statements
Summary:
We first make a pass over the AST to find all the named functions that contains the extracted statements. Then in the second pass, for each function, we insert the `newFunction` declaration to the end of the function body block.

The name of the refactors will be `Extract to inner function in function '%s'`, which matches the current TypeScript language service behavior.

Reviewed By: vrama628

Differential Revision: D29050296

fbshipit-source-id: 5f95415a52167be424c1186d958f2146692406f6