Flow codemod to remove existentials
Summary: This is a codemod to remove the existential types from the codebase and replace them with the inferred types. It works by first trying to infer a type and using that. If there is any type of error, it will replace the existential with `mixed` or `any`. This will depend on whether the `--default-any` flag is passed in.

Reviewed By: panagosg7

Differential Revision: D28855918

fbshipit-source-id: 80ec3fe41ab5ae430520db855e2588c1782d1eea