Add type sig to FindRefsUtils.ast_info tuple
Summary:
The ast_info tuple is used by find-refs and get-def in some places where it
needs to call check_contents. In a follow-up, I am going to add type_sig as a
parameter to that function. In preparation for that, I am adding the type_sig to
this tuple, so we can pass it along.

Reviewed By: panagosg7

Differential Revision: D29033871

fbshipit-source-id: 5c4066e8899ab9fc9c2d2ae1055f2133c09878c2