Change representation of exported locals
Summary:
Exported locals can be identified by their definition location alone, so it's
not necessary to also store the name. This diff uses ALoc.id, which is a special
case of ALoc that is always keyed. During check, we can use the reverse aloc
table to find an ID for an aloc, either concrete or keyed.

In a follow-up, I'm going to remove this set from the file sig and create it
from the type sig during check instead.

Reviewed By: panagosg7

Differential Revision: D29033875

fbshipit-source-id: ac65cc4f49bdcdace9cc7024d090334b5056f8e3