Get exported local information from type sigs
Summary:
Before this diff, the exported locals enviroment was stored in the file sig.
Instead of storing it there, we can derive this information from the type sig
during check.

This saves a small amount of space in shared memory, but more importantly it
removes a dependency on the file sig data, which I'm hoping to eliminate
entirely.

Reviewed By: dsainati1

Differential Revision: D29033870

fbshipit-source-id: c72da948931bb2f87d298244bd1be95e0ebc9e56