Don't use file sig "with exports info" from insert type code action
Summary:
Right now file sig and type sig have a lot of overlap, and I'd like to
consolidate them. The "exports info" data has only a few dependencies. It was
originally introduced for types-first 1.0, but a few other places ended up using
it. This is one of those places.

Instead, we can use type sigs to implement the code here. IMO the resulting code
is even quite a bit simpler.

Reviewed By: vrama628

Differential Revision: D29053511

fbshipit-source-id: 976958e71d79b2ba1ef29d194921d909a7ec0fc1