SharedMem.LocalCache does not need to mutate invalidate_callback_list
Summary:
I want to extract this module so I can invert the control here. Instead of
SharedMem.WithCache, which puts a cache inside the SharedMem API, I want to have
a cache on the outside, wrapping SharedMem calls instead.

This line introduces a dependency from the LocalCache implementation to the
global state of the SharedMem module.

The original motivation for this callback list is that we can clear any caches
associated with shared memory by calling all the invalidation callbacks stored
in the global ref. This seems generally useful, so why is removing this code
okay?

1. These callbacks are also added when we instantiate the WithCache module, so
this code is actually duplicating logic performed elsewhere. It's harmless to
add the callbacks multiple times, but unnecessary.

2. We never actually call the invalidation callbacks anywhere. I could have
deleted the global ref as well, but I'm leaving it alone in this diff to keep
this change minimal.

Reviewed By: vrama628

Differential Revision: D28463572

fbshipit-source-id: cf67401f7bb2a701c90df8f185fe5d9f875a8815