Pass around tables as parameters
Summary:
Before this diff, the type sig parser stored its various tables in the scope
data structure except for locs. I did it this way originally out of laziness,
because I didn't want to thread another parameter through the visitor.

This diff passes a record of tables through the visitor and removes the tables
from the scope data structure. This diff touches a lot of lines, but the changes
are mechanical.

My motivation here is mostly clarity, since there was no good reason for the
tables to be in the scope chain.

Reviewed By: panagosg7

Differential Revision: D29053166

fbshipit-source-id: b2fc593c0b92f3bab1ad183f5796f343a2986060