More precise tracking of existential instantiation for "remove existential" codemod
Summary:
This change aims at getting slightly better results for the "remove existentials" codemod
for the case where the existential is under a polymorphic constructor:
```
class D extends C<*> {
    m() { this.f = ""; }
}
```
The way Flow handles this case is by instantiating a fresh type variable on every reference
to `this.f`.

This diff captures these instantiations and uses them in the codemod.

Of course, we won't be able to always replace the `*` with a type-checking type, since this
feature is fundamentally broken:
```
class D extends C<*> {
    m() {
      this.f = "";
      this.f = 1;
    }
}
```
(Flow is okay with this but the only type that type-check is `empty`.)

In cases like this, the type we were inferring was effectively `any` and the codemod will
most likely fail, and add `any`.

Reviewed By: samwgoldman

Differential Revision: D29035450

fbshipit-source-id: 1aa064c8a2fcd33fa5236de108bac65412bf64ad